swap out libnih and libnih-dbus for glib
The motivation is to make threading possible, to hopefully greatly
speed up systemd startup inside containers.

This required converting all the nih-ified memory tracking.  (Some
of this can probably be done smarter in a more glib-friendly way, i.e.
avoiding doing a glib string alloc followed by strdup followed by
freeing the glib string)

The way the dbus connection is now handled could probably be
improved.  One question is how fuse deals with its threads.
getpid() shows clearly that threads are re-used.  Perhaps
we should keep a dbus connection open as long as a thread is
running?  Share (mutexed) one connection for the whole lxcfs?
Right now we keep a connection open for the duration of a
fuse_* method (when needed).  It is opened on the first
cgcall() that needs it, and closed by hand at end of the fn.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>